+++ src/common/index.js
@@ -11,6 +11,7 @@ export const PLATFORM_AIQICHA = "AIQICHA";
 
 export const TAG_RUOBILIN_BLACK_LIST = "若比邻黑名单";
 export const TAG_IT_BLACK_LIST = "互联网企业黑名单";
+export const TAG_IT_BLACK_LIST_2 = "IT黑名单";
 
 export function getUrlByTagAndCompanyName(tagName, companyName) {
     const decode = encodeURIComponent(companyName);
@@ -18,6 +19,8 @@ export function getUrlByTagAndCompanyName(tagName, companyName) {
         return `https://kjxb.org/?s=${decode}&post_type=question`;
     } else if (tagName == TAG_IT_BLACK_LIST) {
         return `https://job.me88.top/index.php/search/=${decode}`;
+    } else if (tagName == TAG_IT_BLACK_LIST_2) {
+        return `http://www.blackdir.com/?search=${decode}`;
     } else {
         return null;
     }